requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.49
psycopg2-binary==2.9.10
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.8.0
websockets==12.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
